refactor: move loadModels in lib
